Create README.md
Introduction
Football. Pigskin. Soccer. Beach ball. Whatever you call it, American football, the activity that started as a simple game with an oblong ball, has culminated into a billion dollar industry. Not only is this game popular among youth, but the college and professional level of play has become an integral part of American culture. Attending games put on by the National Football League (“NFL”) is a beloved pastime for many in the United States. The NFL was founded 100 years ago, in 1920, and is headquartered in New York City. Players, coaches, staff, aspiring young athletes, and devout fans gather every Thursday, Sunday, and sometimes even Monday throughout the fall to share their love for the game. The American public spends millions of dollars on this industry alone each year from tailgating to game day tickets to team paraphernalia. From a business perspective, the NFL has branded itself as reputable, loveable, and reliable. There is not much that stops these games, and the team owners have ensured that their fans get an optimal game-watching experience through a variety of different business expansion techniques. Some of these include building new stadiums, relocating a team, or just aggressive marketing techniques. For example, the Oakland Raiders are currently in the process of moving to Los Angeles–a move that seems ridiculous to committed locals from Oakland. However, it was executed for the benefit of the league, the financial benefit that is. In 2019 alone, the NFL generated over $8.1 billion in revenue coming out to about $225 million per team. Even more, many Americans spend hours watching the games and engaging with sports broadcasts to learn more about the players. Since the NFL is so influential in the lives of many Americans, we decided to examine more about the league to learn more about just how many people are engaging in this pastime, whether there are any trends in this behavior, whether these trends have changed over time, and more.


The Dataset
The dataset we will use to explore more about the NFL contains information on 32 NFL teams throughout several years. We compiled three different datasets on standings, attendance, and individual game data. In doing so, we were able to conduct analysis on several topics of interest. Our motivation for these analyses is to explore potential relationships between popularity and winning by examining attendance, win/loss records, and the level of difficulty of a given season.


Investigation
The NFL accumulates thousands of statistics on both teams and individuals each year. We will focus on team statistics in order to make better generalizations and understand this topic on a broader scale. For instance, teams are often given a rank by their “strength of schedule.” Teams are divided into conference by their location in the league. In addition, a schedule is made well in advance which means teams have no control over who they play and when. However, they can expect to play similar teams from year to year up until playoffs. Teams that have a lower “strength of schedule” are believed to have played “easier” teams in terms of offensive and defensive rankings. These teams are expected to make it further into the season. This type of data, along with many other statistics, helps us understand how the NFL operates and allows us to gain insight about fans and what they like.


To fine focus his article, we will utilize the five most popular teams in 2020. Popularity was measured in terms of three main variables: Facebook likes, average cost of ticket, and average home attendance. The following were the most popular teams, as of January 2020.
- Green Bay Packers
- Denver Broncos
- New England Patriots
- Dallas Cowboys
- Pittsburgh Steelers

Analysis
Library Dependencies: tidyverse, ggplot2, infer, gganiamte, gifski, png, knitr, prettydoc